1、将扣车表与跳停表合并到一个表，并拆分JSON字段 2、将CLient2serJsonCommand表的json字段拆分出来，添加接收、发送时间等字段 3、添加workstation字段 4、修改版本号：0.5.4.1 	modified:   pom.xml 	modified:   src/main/java/com/byd/ats/controller/PlatformDetainStateController.java 	modified:   src/main/java/com/byd/ats/controller/SkipStationStateController.java 	modified:   src/main/java/com/byd/ats/entity/CLient2serJsonCommand.java 	new file:   src/main/java/com/byd/ats/entity/CLient2serJsonCommandHistory.java 	modified:   src/main/java/com/byd/ats/entity/Client2serCommand.java 	modified:   src/main/java/com/byd/ats/entity/Cu2AtsCiFeed.java 	modified:   src/main/java/com/byd/ats/entity/PlatformState.java 	modified:   src/main/java/com/byd/ats/entity/Ret2ClientResult.java 	modified:   src/main/java/com/byd/ats/entity/SkipStationState.java 	new file:   src/main/java/com/byd/ats/entity/SkipStationState2Cli.java 	modified:   src/main/java/com/byd/ats/rabbitmq/Tut5Receiver.java 	new file:   src/main/java/com/byd/ats/service/Client2serJsonCommandHistoryRepository.java 	modified:   src/main/java/com/byd/ats/service/SkipStationStateRepository.java 	modified:   src/main/java/com/byd/ats/service/SkipStationStateService.java 	modified:   src/main/java/com/byd/ats/service/SkipStationStateServiceImpl.java

+++ src/main/java/com/byd/ats/entity/Cu2AtsCiFeed.java
@@ -15,7 +15,7 @@ public class Cu2AtsCiFeed {
 */
 	private int feed_type;//反馈状态类型
 	private int feed_status;//执行状态
-	private int feed_id;//进路ID/区段ID/道岔ID
+	private long feed_id;//进路ID/区段ID/道岔ID
 	private int feed_time;//计时秒数
 	private long com_serial_num; //序列号
 	public int getFeed_type() {
@@ -30,10 +30,10 @@ public class Cu2AtsCiFeed {
 	public void setFeed_status(int feed_status) {
 		this.feed_status = feed_status;
 	}
-	public int getFeed_id() {
+	public long getFeed_id() {
 		return feed_id;
 	}
-	public void setFeed_id(int feed_id) {
+	public void setFeed_id(long feed_id) {
 		this.feed_id = feed_id;
 	}
 	public int getFeed_time() {
